Reorganize model code.
